# ENH: improve performance of finfo and _commonType (version 2)