--- a/numpy/core/getlimits.py
+++ b/numpy/core/getlimits.py
@@ -469,8 +469,6 @@ class finfo:
 
     """
 
-    _finfo_cache = {}
-
     @lru_cache
     def __new__(cls, dtype):