--- a/numpy/core/tests/test_deprecations.py
+++ b/numpy/core/tests/test_deprecations.py
@@ -900,4 +900,5 @@ class TestRemovedGlobals:
 class TestDeprecatedFinfo(_DeprecationTestCase):
     # Deprecated in NumPy 1.25, 2023-01-16
     def test_deprecated_none(self):
+        np.finfo.cache_clear() # the deprecation warning is only generated once
         self.assert_deprecated(np.finfo, args=(None,))