* Improve performance of `finfo` by making use of an `lru_cache`. For `np.complex128` this makes a large difference.
* Improve performance of  `_commonType` by re-using the expression for `a.dtype.type` and eliminating variables

This is a variation of #23088.

**Notes about the `lru_cache`**

The performance of `finfo` is improved by using an lru_cache

- The behaviour of the new implementation is subtly different. Before we would have `id(finfo(np.float_))==id(finfo(np.complex_))`. In the new implementation this is no longer guaranteed, but the properties of
`finfo(np.float_)` and `finfo(np.complex_)` will be the same.

There are some tests for `getlimits.py` that still pass, but perhaps should be modified. E.g.
```
class TestDouble:
    def test_singleton(self):
        ftype = finfo(double)
        ftype2 = finfo(double)
        assert_equal(id(ftype), id(ftype2))
```
        
- The performance is good for the common cases.

```
python -m pyperf timeit -s "import numpy as np; x=np.array([1.,2.]).dtype;" "np.finfo(x)"
python -m pyperf timeit -s "import numpy as np; x=1.1;" "np.finfo(x)"
python -m pyperf timeit -s "import numpy as np; x=np.array([1.,2.], dtype=np.complex_).dtype;" "np.finfo(x)"
```

Main
```
.....................
Mean +- std dev: 272 ns +- 3 ns
.....................
Mean +- std dev: 2.40 us +- 0.06 us
.....................
Mean +- std dev: 272 ns +- 5 ns
```

Pr:

```
.....................
Mean +- std dev: 212 ns +- 2 ns
.....................
Mean +- std dev: 237 ns +- 2 ns
.....................
Mean +- std dev: 211 ns +- 2 ns

```

 There are scenarios where many cache misses occur (for example `np.finfo(x)` is allowed for any float `x`, so something like `[np.finfo(x) for x in np.arange(10000)]` will result in misses), but they are hardly any slower than the old implementation.

- Just adding an `lru_cache` is sufficient for the performance. But since it is no longer required, the `_finfo_cache` is removed. 
- The original implementation contained multiple calls to the cache. They have all been removed to simplify the implementation and avoid any risks of creating circular references.
- The `lru_cache` caches the combination of the `cls` and `dtype` argument. The extra overhead compared to just caching the `dtype` is not very large. A test:
```
%timeit hash((cls,d))
%timeit hash(cls)
%timeit hash(d)
108 ns ± 4.66 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
56.3 ns ± 1.55 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
61.1 ns ± 0.2 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
```
One might eliminate this by changing `finfo` from a class into a function `def finfo(dtype)` that returns the class, but that is not implemented in this PR.

- The deprecation warning for `np.finfo(None)` will only get printed once.
- For integers there is the `iinfo` class. No caching was present there, but performance is improved with a `lru_cache` as well.


<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->