@eendebakpt 2023-01-26
The performance of `iinfo` can be improved by
```
Main: Mean +- std dev: 1.22 us +- 0.07 us
PR: Mean +- std dev: 233 ns +- 12 ns
```
by replacing the `__init__` with
```
    @lru_cache
    def __new__(cls, int_type):
        return object.__new__(cls)._init(int_type)
    
    def _init(self, int_type):
        try:
            self.dtype = numeric.dtype(int_type)
        except TypeError:
            self.dtype = numeric.dtype(type(int_type))
        self.kind = self.dtype.kind
        self.bits = self.dtype.itemsize * 8
        self.key = "%s%d" % (self.kind, self.bits)
        if self.kind not in 'iu':
            raise ValueError("Invalid integer data type %r." % (self.kind,))
        return self
```