@seberg 2023-01-26
You are right, you would at least need to use the `typed=True` lru-cache to ensure that you don't run into problem since scalar instances are currently allowed (for that matter, I wouldn't be surprised if *array*s work).

Maybe just bail on this a bit and keep the dtype/type conversion logic in either case then add the cache (or lru) only then?  Making sure complex ends up in the cache is important, but the other part seems to need a bit of care without input validiation.