@seberg 2023-01-27
You need at least `lru_cache(typed=True)`?  If you are confident about that, I can go along with that.  But please add tests for `float64(1.)` and `float32(1.)` (and maybe even integer 1 for that matter).

But unless the code is getting a bit cleaner, limiting to just the complex performance fix seems also OK.