@eendebakpt 2023-01-27
> You need at least `lru_cache(typed=True)`? If you are confident about that, I can go along with that. But please add tests for `float64(1.)` and `float32(1.)` (and maybe even integer 1 for that matter).
> 
> But unless the code is getting a bit cleaner, limiting to just the complex performance fix seems also OK.

I don't think we need `typed=True`. Why do you think so? 

Your suggestion to move the lru_cache a bit down I worked out in https://github.com/numpy/numpy/compare/main...eendebakpt:numpy:linalg_performance_v3

All three versions have pros and cons, I will post the performance numbers in a bit. But I am not sure it matters a lot. The bad performance for complex numbers I reported was on `np.complex128`, but not on `np.complex128(1.).dtype`, and I suspect the latter one is the more important one.

The fact that `np.finfo(12.)` is allowed makes the implementations a bit harder, but for performance it might not matter. Or are you aware of cases where it does? The case that led me to `finfo` is the `np.linalg.lstsq` which only passes the dtype of a numpy array to finfo.