@seberg 2023-01-27
```
def test():
   assert np.finfo(np.longdouble(1)) is np.finfo(nplongdouble)
   assert np.finfo(np.float64(1)) is np.finfo(np.float64)
   assert np.finfo(np.float32(1)) is np.finfo(np.float32)
   assert np.finfo(np.float16(1)) is np.finfo(np.float16)
   with pytest.raises(TypeError):  # or value
       np.finfo(np.longdouble(1))

```
Same for iinfo, the code here doesn't look like it would pass, both other versions should though.  So either seems fine and I would say pick whichever seems more maintainable/easy to understand (I don't think performance difference will be relevant).