@eendebakpt 2023-01-27
@seberg The condition `assert np.finfo(np.float64(1)) is np.finfo(np.float64)` does indeed not hold for v2, but also not for v3. I think for anything with an `lru_cache` the following will also fail:
```
f1 = np.finfo(np.float64)
# ... do some work that fills the lru cache and drops out f1
....
f2 = np.finfo(np.float64)
assert f1 is f2
```
If this is a requirement, then I will add the tests from your comment above, and keep version 1 as that is the only one that satisfies such requirements. The original [commit](https://github.com/numpy/numpy/commit/e421eb9e41bde5579e097cd7b7ec1b6f1a6b10bb) only contains the text `Impl. finfo as a new-style singleton class.` suggesting this is intended behaviour, but not with a reasoning why.

If we only require `np.finfo(np.float64(1)) == np.finfo(np.float64)` (with `==` comparing the fields of the `finfo` instance), then we can use either v1, v2 or v3 and we can further simplify the implementation of v1.

Based on performance all options seem an improvement over current main (assuming the main performance critical case is `finfo(d)` with `d` are dtype).

<details><summary>Performance tests</summary>
Benchmark
```
echo "Common" 
python -m pyperf timeit -s "import numpy as np; x=np.array([1.,2.]).dtype;" "np.finfo(x)"
python -m pyperf timeit -s "import numpy as np; x=1.1;" "np.finfo(x)"
python -m pyperf timeit -s "import numpy as np; x=np.array([1.,2.], dtype=np.complex_).dtype;" "np.finfo(x)"
python -m pyperf timeit -s "import numpy as np;" "np.finfo(np.complex128)"

echo "hard cases for the lru_cache"
python -m pyperf timeit -s "import numpy as np; x=np.linspace(0,100,3023)" "[np.finfo(v) for v in x]"
python -m pyperf timeit -s "import numpy as np; x=[1.23*v for v in range(3000)]" "[np.finfo(v) for v in x]"

echo "iinfo"
python -m pyperf timeit -s "import numpy as np; x=np.array([1]).dtype;" "v=(np.iinfo(x))"
python -m pyperf timeit -s "import numpy as np; x=np.array([1]).dtype;" "v=(np.iinfo(2))"
```

Main
```
Common
.....................
Mean +- std dev: 470 ns +- 20 ns
.....................
Mean +- std dev: 2.29 us +- 0.06 us
.....................
Mean +- std dev: 465 ns +- 14 ns
.....................
Mean +- std dev: 522 ns +- 4 ns
hard cases for the lru_cache
.....................
Mean +- std dev: 2.41 ms +- 0.08 ms
.....................
Mean +- std dev: 7.58 ms +- 0.12 ms
iinfo
.....................
Mean +- std dev: 1.13 us +- 0.03 us
.....................
Mean +- std dev: 2.82 us +- 0.10 us
``` 

v1
```
Common
.....................
Mean +- std dev: 271 ns +- 2 ns
.....................
Mean +- std dev: 2.39 us +- 0.04 us
.....................
Mean +- std dev: 274 ns +- 13 ns
.....................
Mean +- std dev: 306 ns +- 4 ns
hard cases for the lru_cache
.....................
Mean +- std dev: 2.74 ms +- 0.06 ms
.....................
Mean +- std dev: 8.08 ms +- 0.29 ms
iinfo
.....................
Mean +- std dev: 1.13 us +- 0.04 us
.....................
Mean +- std dev: 2.82 us +- 0.07 us
```

v2
```
Common
.....................
Mean +- std dev: 212 ns +- 2 ns
.....................
Mean +- std dev: 239 ns +- 3 ns
.....................
Mean +- std dev: 213 ns +- 3 ns
.....................
Mean +- std dev: 248 ns +- 3 ns
hard cases for the lru_cache
.....................
Mean +- std dev: 59.9 ms +- 1.7 ms
.....................
Mean +- std dev: 8.74 ms +- 0.18 ms
iinfo
.....................
Mean +- std dev: 216 ns +- 3 ns
.....................
Mean +- std dev: 210 ns +- 3 ns
```

v3
```
Common
.....................
Mean +- std dev: 400 ns +- 6 ns
.....................
Mean +- std dev: 2.50 us +- 0.07 us
.....................
Mean +- std dev: 402 ns +- 9 ns
.....................
Mean +- std dev: 695 ns +- 27 ns
hard cases for the lru_cache
.....................
Mean +- std dev: 3.08 ms +- 0.13 ms
.....................
Mean +- std dev: 8.26 ms +- 0.23 ms
iinfo
.....................
Mean +- std dev: 215 ns +- 7 ns
.....................
Mean +- std dev: 208 ns +- 3 ns

```
</details>